% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozbabynames-package.R
\docType{package}
\name{ozbabynames-package}
\alias{ozbabynames-package}
\title{ozbabynames: Australian Popular Baby Names}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Data on the most popular baby names in Australia.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robjhyndman/ozbabynames}
  \item \url{https://pkg.robjhyndman.com/ozbabynames/}
  \item Report bugs at \url{https://github.com/robjhyndman/ozbabynames/issues}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Mitchell O'Hara-Wild (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})
  \item Jessie Roberts
  \item Nick Tierney (\href{https://orcid.org/0000-0003-1460-8722}{ORCID})
}

Other contributors:
\itemize{
  \item Anna Fergusson (\href{https://orcid.org/0000-0002-1987-8150}{ORCID}) [contributor]
}

}
\keyword{internal}
