% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozbabynames.R
\docType{package}
\name{ozbabynames}
\alias{ozbabynames}
\alias{ozbabynames-package}
\title{ozbabynames: Popular Australian baby names.}
\format{tibble}
\source{
Various state government websites
}
\usage{
ozbabynames
}
\description{
The ozbabynames package provides the data object `ozbabynames` containing
popular Australian baby names by sex, state and year. The coverage is very uneven,
with some states only providing very recent data, and some states only providing the top 50 or 100
names.
}
\examples{
head(ozbabynames)

}
\keyword{datasets}
